requirements.txt
python-dotenv
langchain
langchain-community
langchain-ollama
requests
chromadb
ollama
